+++ src/utils.h
@@ -1,11 +1,12 @@
-#ifndef DYNGEN_UTILS_H
-#define DYNGEN_UTILS_H
+#pragma once
 
 #include <Rcpp.h>
 using namespace Rcpp;
 
+namespace gillespie{
+
 template <typename T>
-T resize(const T& x, int n){
+T resize_vector(const T& x, int n){
   int oldsize = x.size();
   if (n < oldsize) {
     oldsize = n;
@@ -43,24 +44,6 @@ int weighted_sample(const NumericVector& weight) {
   return j;
 }
 
-void fill_nu_vectors(
-    const IntegerMatrix& nu,
-    IntegerVector& nu_row,
-    IntegerVector& nu_effect,
-    bool* nu_single
-) {
-  for (int j = 0; j < nu.ncol() && *nu_single; j++) {
-    for (int i = 0; i < nu.nrow(); i++) {
-      if (nu(i, j) != 0) {
-        if (nu_effect[j] == 0) {
-          nu_effect[j] = nu(i, j);
-          nu_row[j] = i;
-        } else {
-          *nu_single = false;
-        }
-      }
-    }
-  }
-}
+} // namespace gillespie
 
-#endif
+// #endif